fix(RHINENG-12947) close dropdown on outside click
This PR https://github.com/RedHatInsights/vulnerability-ui/pull/2161 implemented a new version of the Publish date filter in the Reports modal. But it introduced a bug. When you opened the Publish date dropdown and then clicked on another dropdown, the Publish date dropdown would stay open.

This PR makes sure the dropdown closes when you click out of it. It also wraps the component in a div with the proper class and styles to add padding.