feat(RHINENG-16846): Add operating systems report card
This PR adds a new card to the Reports page. It is hiddent behind a
feature flag, `vulnerability.exposure-by-os-minor.enabled`. To test,
simply go to Vulnerability -> Reports and check for the existence of the
OS report card when the flag is flipped on or off. The link on the card
should not be functional.

This PR also refactors the rendering of the cards as well as removes a
redundant useFeatureFlag hook. It also removes some intl usage.