feat(RHINENG-17936): update exposure selection
This PR updates the way we make OS version and lifecycle selections for the PDF. We want to use the `selectedOsVersions` array as a reference for the pdf to get only the data it needs. So if a user select RHEL 8.8, it will default to only the `minor` lifecycle data.
```
[

    {

        8.8: {

           minor:  { ...data }

        }

    }

]
```

And, if a user changes the lifecycle selection to Extended Support:
```
[

    {

        8.8: {

           eus: { ...data }

        }

    }

]
```