feat(RHINENG-18786): Add system count to pdf
This PR adds a system count to the RHEL major version tables in the
"Report by operating system versions" report on the Reports page.

To see, go to Vulnerability -> Reports and click the "Report by
operating system versions" card link. Select up to 5 os versions and
Export PDF. The rendered PDF should show the total number of systems for
your selected RHEL major versions.

So, if you selected RHEL 9.3 and 9.4, it will provide you with a count
of ALL systems using RHEL 9.

In addition, the bar chart on the last page of the PDF includes a list
of all of the RHEL minor versions underneath the title.